# feat(cli): enhance archive summary display with terminal compatibility